Rename fact docker_version to ceph_docker_version
The name docker_version is very generic and is also used by other
roles. As a result, there may be name conflicts. To avoid this a
ceph_ prefix should be used for this fact. Since it is an internal
fact renaming is not a problem.

(cherry picked from commit 50a848dc408a35c02b934bfe1511cd8aaee259be)
Signed-off-by: Sébastien Han <seb@redhat.com>