Cleanup fsid generation
The current approach makes fsid generation complicated by utilising a
file and fetch directory. If you delegate a set_fact to localhost it
will set the fact for all other hosts.

In this way we can set the fsid based on the running cluster (if it's
running), or generate it as before - but only perform this once and all
hosts can consume that setting.

Additionally - to allow this to run as part of the rolling updates, we
can run the task to check if ceph is running against either mons[0] or
mons[1] if we are currently using mons[0].