defaults: run pre-checks before each role is played
This moves all the checks that are currently split into each role to
`ceph-defaults` so we can fail early in the playbook if something is
wrong. Indeed, there is no point to start deploying the cluster if it's
going to fail anyway because of a wrong variable definition or anything
similar.

Signed-off-by: Guillaume Abrioux <gabrioux@redhat.com>