defaults: remove `run_once: true` when creating fetch_directory
because of `serial: 1`, it can be an issue when the playbook is being
run on client nodes.
Since the refact of `ceph-client` we skip the role `ceph-defaults` on
every node except the first client node, it means that the task is not
going to be played because of `run_once: true`.

Signed-off-by: Guillaume Abrioux <gabrioux@redhat.com>
(cherry picked from commit dcf6a246a44e82ebf3092c696b9f86d755157e71)
Signed-off-by: Sébastien Han <seb@redhat.com>