choose_disk: check if the device has a partition table
We shouldn't fail if the device does not have any partition table. So if
we find it does not have any we return the device as usable.

If the device has one, choose_disk moves to the next task and tries to
look what are the partitions.

Typically loop device by default won't have any partition table and raw
drives too. So we should include them in our selection.

Signed-off-by: Sébastien Han <seb@redhat.com>