Install Apt packages as a list
We can reduce the number of tasks required, and install all packages in
one go, rather than looping through them or installing them as
individual tasks. This will be more efficient since a list can be passed
to the apt install process.

This PR adds a list creation based on the conditionals used, and then
installs the packages in one go.