Install packages as a list
To make the package installation more efficient we should install
packages as a list rather than as individual tasks or using a
"with_items" loop. The package managers can handle a list passed to them
to install in one go.

In order to reduce the need for a task to sort the packages we can
create a var that will filter the packages and produce the list of
packages to install.