choose_disk: Sanitizing input
The actual code was failing with input like :
    rotational:
        eq : '0'

The 0 value written in string was preventing the match.
The reason was the conversion to int was only done for the 'size' feature.

This patch offer to make a generic sanitizing of inputs by checking
every right operand and turn then into float if possible or let them
untouched. For that, we reuse to_bytes() which is capable of
understanding specific units.
As we do convert everything in float, can_this_be_an_int() is then removed.

If the right operand is not a dict, it means we have an input like:
    rotational: '0'.

In this particular case, a default 'eq' operand is injected. This will
make user's life easier as 'eq' is usually the default expectation.

Signed-off-by: Erwan Velu <erwan@redhat.com>