choose_disk: Use parent device in legacy matching
When matching legacy devices, the lookup block device is reported as the
full block device like /dev/sdb.
At the initial creation of the OSD, that's perfect. But once ceph_disk
have prepare the OSD, that's a different game.

ceph-ansible still lookup for /dev/sdb but its /dev/sdb1 which have the
expected metadata. As a result, the matching is failing like in :

    INFO Detecting free devices with fsid=b4c56e01-ab71-4b24-93d0-adafd139847f
    INFO Ignoring       dm-0 : device is a DM
    INFO Ignoring       dm-1 : device is a DM
    INFO Adding         sda1 : Ceph disk detected (data)
    INFO Adding         sdb1 : Ceph disk detected (data)
    INFO Adding         sdc1 : Ceph disk detected (journal)
    INFO Adding         sdc2 : Ceph disk detected (journal)
    INFO Ignoring        vda : device has existing partitions
    INFO Legacy syntax
    INFO devices : ['/dev/sdb', '/dev/sda']
    INFO dedicated_devices : ['/dev/sdc']
    INFO Looking for {'journal_0_000': {'ceph_type': 'journal', 'bdev': '/dev/sdc'}, 'data_0_000': {'ceph_type': 'data', 'bdev': '/dev/sdb'}, 'data_1_000': {'ceph_type': 'data', 'bdev': '/dev/sda'}}
    INFO Looking for matches
    INFO Inspecting data_0_000
    INFO  sdb1 : no match /dev/sdb1 {'eq': '/dev/sdb'}
    INFO        sdb1 no devices matched

This patch is about detecting this situation and does the following:
- if we run in legacy
- and we search of a bdev
- we replace the partition by its parent for the match (/dev/sdb1 become /dev/sdb)

Thanks to this trick, the matching now pass in that case.

Signed-off-by: Erwan Velu <erwan@redhat.com>